🐛 Threading ReloadSocketServer
Necessary to handle the fact that, during reload, the "save-output" connection
in the top tuick extends until the end of the build, and nested tuick format
commands send their own save-output message within the lifetime of the top
save-output.

Code was already thread-safe.

Should probably be rewritten in async style.