🚧♻️🐛 Restructure raw output collection to support nested tuick. NEED TEST
Add begin-output, end-output messages to mark the start and end of a raw dump.

Needs integration tests for the "tuick make" scenario with nested tuick.
Tests for initial load and reload. Must verify that the final output is the
concat of all raw outputs, in particular the top save-output commands are
interspersed with bottom save-output commands.