🚧💄 Nested tuick does not output empty nested blocks marker
The intent was to combine the informational lines before and after a nested
tuick with no output. So "make foo bar" may show "Doing foo" and "Doing bar",
and if nested tuicks in the foo recipe produce no output, these two lines
should be combined in one block for faster navigation.